Extract chat window bindings
