+++ tests/test-custom-personality.js
@@ -207,7 +207,7 @@ assert('sendChatMessage handles AbortError', sendSrc2.includes('AbortError'));
 // ── 19. Stop button CSS ──
 console.log('19. Stop button CSS');
 const css = read('styles.css');
-const chatSrc = read('js/chat.js');
+const chatWindowBindingsSrc = read('js/chat-window-bindings.js');
 const chatDiscussionSrc = read('js/chat-discussion.js');
 assert('CSS has .chat-send-btn.streaming', css.includes('.chat-send-btn.streaming'));
 assert('CSS has .chat-stopped-note', css.includes('.chat-stopped-note'));
@@ -263,7 +263,8 @@ assert('endDiscussion cleans up state', endSrc.includes('cleanupDiscussionState'
 assert('endDiscussion marks discussion ended', endSrc.includes('markEnded: true'));
 assert('chat restores discussion prompt without thread metadata', chatDiscussionSrc.includes('restoreDiscussionContinuePrompt') && chatDiscussionSrc.includes('collectDiscussionPersonas()'));
 assert('manual messages in active discussion suppress duplicate auto prompt', chatDiscussionSrc.includes('suppressAutoMsg: true'));
-assert('chat.js configures discussion callbacks', chatSrc.includes('configureChatDiscussion') && chatSrc.includes('setChatAbortController'));
+assert('chat window bindings configure discussion callbacks',
+  chatWindowBindingsSrc.includes('configureChatDiscussion') && chatWindowBindingsSrc.includes('setChatAbortController'));
 assert('endDiscussion restores personality', endSrc.includes('currentChatPersonality'));
 
 // ── 27. Steer input CSS ──
